+++ benches/benches/block_target_gas_set/mod.rs
@@ -1,3 +1,8 @@
+use fuel_core_types::fuel_tx::{
+    DependentCost,
+    GasCostsValues,
+};
+
 pub mod alu;
 
 pub mod crypto;
@@ -7,3 +12,5 @@ pub mod flow;
 pub mod contract;
 
 pub mod memory;
+
+pub mod default_gas_costs;