+++ crates/fuel-core/src/database.rs
@@ -21,6 +21,7 @@ use fuel_core_storage::{
     Result as StorageResult,
 };
 use fuel_core_types::fuel_types::BlockHeight;
+
 use itertools::Itertools;
 use serde::{
     de::DeserializeOwned,